refactor(frontend): split lecture studio editor controls/options
